Rename ListPropertyFactory.CodeGenerator -> ListProperty
ListPropertyFactory is now ListProperty.Factory
